Clientes, ClientesRepository, listado y MainController actualizados

+++ src/main/java/com/alfonsotienda/holaspring/model/ClientesRepository.java
@@ -3,13 +3,9 @@ package com.alfonsotienda.holaspring.model;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-/**
- * ClientesRepository.findAll() mirar los demás después del punto
- * Devuelve Iterable<T> y findAll() Returns all instances of the type.
- */
 
 @Repository
 public interface ClientesRepository
-    extends CrudRepository<Clientes, Long>{
+    extends CrudRepository<Clientes, Integer>{
 
 }
